(FACT-1155) Use stat over access for consistency.
On some platforms (looking at you, Solaris), `access` returns 0 when run
as root even if a file has no executable bits set.  This results in unit
test failures for libfacter on Solaris because files that should not be
considered executable are being treated as executable.

The fix is to use `stat` to check for access.  This requires to check
for user, group, and other executable bits in order.