(FACT-1155) Fix ruby unit test to not rely on behavior of sh.
A ruby unit test was relying on the behavior of `sh` for handling
stdout/stderr redirection of failed commands. On some platforms (i.e.
Solaris), `sh` is not redirecting the output of a command that doesn't
exist, resulting in an unexpected result for the fact's value.

The fix is to use a command that gives output on stdout, but ultimately
returns a non-zero exit code.