(maint) Treat zero processor counts as missing facts
Previously, the processor resolver would faithfully report a value
of 0 for the logical or physical processor counts. However, that is
counterintuitive (since clearly no system ever actually has zero
processors) and is inconsistent with handling of other zero values
both in the processor facts (e.g. a speed of zero is treated as a
missing fact) and with other facts (e.g. zero network interfaces
is treated as a missing fact).

So this simply treats a value of zero for the logical and physical
counts as a missing fact.

This is of real-world interest for platforms, like OpenBSD, which
don't have a way of reporting the physical processor count.