(FACT-1152) Fix Linux partitions fact to enumerate devices like 2.x.
Facter 2.x looked through the /sys/block structure to detect partitions
to populate the fact.  3.x changed this behavior to use libblkid to
enumerate the devices.

Using libblkid to enumerate the devices is problematic as probing
requires root access and if no cache file is present, normal users won't
the partitions populated correctly.

This changes the probing to the 2.x with additional support for
displaying mapped and loop devices.

Now libblkid is only used to query the device attributes, meaning if
libblkid support isn't enabled, the partitions will just show their size
and mount points.

This also fixes the size detection for mapped devices as the opening of
the size file was being done with the mapping name and not the
underlying block device name.  That resulted in some partitions showing
as having 0 bytes.  Now the correct name is used under /sys/block.

This also adds a "backing_file" attribute for partitions that are loop
devices.