(FACT-970) Implement processor resolver for AIX
This uses the AIX [object data manager][odm] to query the system for
all available processors and hardware threads.

Since I'm not sure if other resolvers will use the ODM, it currently
uses the APIs directly. If we see more use of it in other resolvers we
should consider abstracting things, as there is some serious weirdness
with this API.

[odm]: https://www-01.ibm.com/support/knowledgecenter/ssw_aix_71/com.ibm.aix.genprogc/odm.htm