(FACT-970) remove physical_count from AIX processor resolver
AIX reports information on processor *cores*, not on physical
processors. physical_count would be wrong in the case of a dual- or
quad-core processor.

We still report the overall count and the model array as the total
number of SMT threads. This matches Linux facter 3, but is different
from AIX facter 2 (which reported only the number of cores)