(FACT-1158) Fix expanding commands with spaces in the path.
On Windows, expanding commands like "git" which are on the path under the
program files directory that contains a space, resulted in the command not
being properly quoted.  The result was a failure to execute the command
because the command shell was not properly given a quoted path.

Two fixes are needed: execution::expand_command should properly quote the
expanded command if it contains a space and Windows execution should
respect an option to "preserve" the arguments as given because the Ruby
execution functions pass the command lines themselves (user's
responsibility to quote arguments).