(FACT-1167) Fix Solaris networking resolver not resolving primary interface.
The Solaris networking resolver was missing the fallback logic to assign the
primary interface to the first interface that contains a valid address.

This logic existed in the BSD resolver, but it makes more sense to move it into
the base resolver itself so that the fallback logic is shared between all
platforms.