(maint) Refine debug messages for Solaris processors fact
The prior debug messages conveyed the impression that processor facts
would not resolve, when the fallback would likely still resolve them.
Change the messages to be more specific about the failure.

Also change behavior to be clearer, that we should fallback to calling
kstat only if there was an exception trying to use the kstat api.