(FACT-1243) Add Scientific/Oracle to el acceptance
This commit adds Scientific and Oracle to the accepted pattern
of os.name values for the `acceptance/test/facts/el.rb` test.

Prior to this commit, this test would fail On Scientific Linux
and Oracle Linux due to the fact that the `os.name` value did
not match the available values.