(FACT-1238) Respect fully-qualified hostnames if specified.
Facter 2.x would respect any hostname that was already fully-qualified.  A
hostname of 'foo.bar.com' would be split into a hostname of 'foo' and a domain
of 'bar.com'. No further attempt to resolve the FQDN would take place.

Facter 3 always used fallback logic of attempting to resolve the hostname to a
FQDN and looking through `resolv.conf` for domains.  This should only occur if
the system hostname is not already fully-qualified.

This is to say, if the users have configured their machine hostnames to be
fully-qualified, Facter should respect that.