(FACT-780) Prevent Puppet from altering Facter logging with -p.
When the `-p` option is given to Facter, Facter loads and initializes Puppet.
However, part of the Puppet initialization is to alter Facter's logging
behavior to better integrate it with Puppet's output.

When Puppet is loaded by Facter, this is undesirable behavior as by default
this will cause no further output from Facter other than printing the facts.

To fix this, this change only defines the methods that Puppet consumes to alter
Facter's logging behavior if the `-p` option was not specified.  Puppet already
checks for these methods being present before consuming them.