(FACT-1254) Allow dot syntax to query ruby facts
Previously, the dot syntax could not query ruby facts, since they were
opaque to the facter collection.

This commit updates the collection code to recognize when it has
encountered a ruby value during dot lookup, and passes the value and
the remaining query segments to the facter::ruby code to finish the
query.