(FACT-380) Use the interface of the default route for network facts
Previously, Facter would use whichever non-loopback device came first
in the ifconfig output for the networking facts. This is not always
what most people would consider the "ipaddress of a node", since
things like virtual network devices can show up in this list and might
sort before physical interfaces.

This changes the code to check the routing table for the default
interface, and uses the values from that interface.