(FACT-1282) Populate interface bindings from routing table
Previously, we only populated interface bindings directly from the IPs
that were set on that interface. It's possible, though, that an
interface doesn't have a direct IP binding, but instead only has a
'source' parameter on a route that sends outgoing traffic through that
interface. This happens when incoming and outgoing traffic is split
between two interfaces.

This patch queries the source parameter of routes, and treats them as
additional bindings on the interfaces referenced in the routing table.