(maint) Skip check for locale warning on OS X
i18n work in Leatherman 0.4.0 introduced using Boost.Locale to generate
locales on platforms besides Windows. For OS X, this has a side-effect
that an exception is no longer thrown if a bad locale is set. Add OS X
to the list of platforms we don't check for a warning when a bad locale
is set.

[skip ci]