(FACT-1362) Add platform identification for HuaweiOS
HuaweiOS runs as an LXC container based on Debian 8 with some
modifications (for example, it uses Debian sysv instead of
systemd). It still uses apt and dpkg for package management.

The firmware version is a long string stored in
/etc/huawei-release which does not follow semver, but looks
like "V100R006C00SPC300B188", so we just pass along this
version information without further parsing.

So the os.name we're using is "HuaweiOS", and os.family is
"Debian"