(FACT-1389) Fix facter.rb regression
The work for FACT-1294 caused a regression when running Facter from MCO
(as a service), where we try to load facter.rb but Puppet
Labs/facter/bin isn't on the PATH. Previously facter.rb ensured it was
added to the path, but the change in FACT-1294 accidentally dropped that
(as part of an assumption that the next release would be with new MSI
packaging that put ruby.exe and libfacter in the same directory).

Restore setting PATH as part of loading libfacter. Also adds an
acceptance test to avoid further regressions.