(FACT-1385) Support negative fixnum values with Ruby 2.2
Because the method `rb_num2long` was changed in Ruby 2.2. to return
a long, some unit tests were failing in Ruby 2.2. on windows because
they were expecting it to return a 64-bit type. Update facter to use
`num2size_t` and `rb_num2ll` methods instead in order to fix this
issue.

This fix also relies on an update to leatherman which adds the
ability to recognize bignums to the ruby API.