(FACT-1364) Support bignum fact values
Prior to this commit, when printing out fact values, facter only
supported numbers it could identify as a fixednum or a double. This
meant that anything that was a bignum would not be printed. Update
the write and to_json methods so that they support bignums.

This fix also relies on an update to leatherman which adds the
ability to recognize bignums to the ruby API.