(FACT-1373) Prevent fact calls from external facts
Prior to this commit, it was possible to make a facter call from
within an external fact. This would result in a fork bomb where
facter would be constantly calling into itself. In order to prevent
this, set an environment variable that allows us to ensure facter
is not already running when we make the call to evaluate external
facts.