(PE-14501) add superuser flag to the identity fact
This commit adds the `superuser` flag as another attribute of the `identity`
structured fact.
It is a boolean flag which is set to `true` if the facter process runs with
UID of 0 on *NIX systems or with the privileges of the local Administrators
group on Windows. Under any other circumstances the flag is set to `false`.