(FACT-1424) Remove Processor architecture test on platform name
The Processor architecture tests in facts/windows.rb determine the type
by the agent platform however processor type should be determinded by
the hardware i.e. It is possible to have a 32bit OS on a 64bit Processor.
This commit changes the test to accept either x64 or x86 as acceptable
values.