(FACT-1438) Add s390x support for acceptance
This commit modifies the el acceptance tests to accommodate
support for the s390x architecture on the zlinux platform.

Since test instances on s390x do not use vmpooler, we cannot use
the beaker `host['template']` value to determine the expected OS.
Therefore, this value has been capturing the contents of the
`/etc/*-release` file on the el based system.