(FACT-1244) Correctly report Xen Dom0 vs DomU
Previously, our virt-what reader would return the first line of output
it found. For cases where the machine is Xen, however, this first line
is simple "xen", with Dom0 or DomU information being provided
after. In that case we want to skip the first line and move on to the
detailed hypervisor info.