(FACT-1464) Fix duplicate Ruby GC registration in module::create_fact
Prior to this fix, a single fact object could in some circumstances be registered twice with Ruby GC. While populating the collection, the facts() method can initiate a recursive call to create_facts(), resulting in two calls to _facts.insert() with the same fact_name. The second call to _facts.insert() will return an iterator to the first object inserted and the subsequent call to rb_gc_register_address() will re-register this object.