(FACT-1460) Load global settings from config file when loading Facter from Ruby
This commit includes a refactor which builds config file access out into
a utility file. It also enables config file loading when Facter is
called from Ruby. For this mode, only two of the currently implemented
settings are used: external-dir and custom-dir, which allow the user to
specify additional directories for custom and external facts. These are
added to any paths specified by Puppet when running Facter from within
Puppet.