(FACT-1356) Allow executable facts to output Yaml/Json
Previously, executable facts could only return simple key/value
pairs. This code now attempts to parse executable output as Yaml, and
only falls back to the key/value parsing if we do not get a useful
result. Because Yaml is a superset of Json, we do not have to
seperately try to parse Json output.