(FACT-1518) Add quotes to ttl keys so hocon lookup works correctly
When the key string for each ttl object is passed directly to
config::get_duration, cpp-hocon will parse it as an unquoted string and
choke on special symbols. This adds triple quotes, to treat the string as a
single element with all special symbols ignored.

Note: this disables escape sequences and path splitting on '.'