(FACT-1441) Add basic "cloud" fact
Since there currently exists no fact that can identify if a machine is running
inside Azure, we need to add a more robust cloud fact. This fact will be
differentiated from the existing `is_virtual` and `virtual` facts in that
it will allow extra information due to its structure.

This step adds the `cloud` fact and checks for azure on linux.