(FACT-1561) Use top namespace when extending TestCase
A recent change in Beaker renamed one of the module names from
"FacterHelpers" to simply "Facter". This exposed a flaw in the way tests
in this repo were extending the TestCase object. When not specifying the
top namespace, the `extend` method introspects the current module first
before looking out to the `$LOAD_PATH`. Now that `Facter` is a module
included in the DSL, attempting to execute `extend Facter::Lib` will
result in ruby looking into `Beaker::DSL::Facter::Lib`.

This change prepends a `::` to all appropriate module names in testing,
so that ruby knows to look for these helper libraries at the top level
namespace. Fixing the tests here will make them more resilient to future
beaker changes.