(maint) Do not rely on external networking for WSASetup validation
For FACT-1551, we added a test to ensure we had called WSASetup before
trying to resolve Ruby facts. As originally implemented, this test
required public network access and a functional external resource.

Most of this is not necessary to validate that WSA has been setup, and
makes the test more fragile. Instead, we can simply ensure that
creating and closing a socket proceeds without error.