(FACT-1583) Output exception message when test fails
For reasons unknown, the custom fact that relies on networking sometimes
fails on Fedora 24, but the test output only says that "" doesn't equal
"Yay".

If an exception occurs, output the message as the fact resolution, which
won't equal "Yay". That way we can debug the issue if/when it occurs
again.