(FACT-1595) remove the file ttls_cached_facts_create_json_cache_file.rb that got renamed incorrectly
In the 2 previous commits for FACT-1595, the intent was to rename the ttls.rb test into
ttls_cached_facts_creates_json_cache_file.rb (with the "creates"). However when making
updates to the commits and their interactions between each other the file was renamed
to ttls_cached_facts_create_json_cache_file.rb (with out the 's' for "create").
We don't need this multi-test, test file as its covered by these test files:

  ttls_cached_facts_creates_json_cache_file.rb
  ttls_cached_facts_expire_facts_do_not_read_the_old_cached_value.rb
  ttls_cached_facts_expire_facts_refresh_the_cached_value.rb
  ttls_cached_facts_read_from_the_cached_value.rb
  ttls_cached_facts_that_are_corrupt_are_refreshed.rb
  ttls_cached_facts_that_are_empty_return_an_empty_value.rb
  ttls_puppet_facts_creates_json_for_cached_facts.rb
  ttls_puppet_facts_honors_cached_facts.rb