(FACT-1675) fix virtual, and is_virtual facts for OpenBSD VMs
Without the patch, OpenBSD VMs running within OpenBSD's hypervisor
vmm will wrongly report the fact virtual -> physical, and is_virtual -> false
With the patch, the following will be reported: virtual -> vmm, is_virtual -> true

It's unfortunate, but hw.product doesn't return anything in such
cases, therefore in openbsd/dmi_resolver, set the make use of
hw.vendor to set the product name.
The rest is rather mechanic, add a string for the hypervisor to facts/vm
as well as a mapping from the sysctl value to the string to
virtualization_resolver.