(packaging) Stop building all the things
We are only making gem and tar archives of facter 2 available, but with
these settings, we were building rpms, dmgs, debs, etc. That wasn't
great, because that automation required old systems that had either been
abandoned or decommissioned. This PR removes that to allow us all to
move on with our lives.