(FACT-1778) Attempt to canonicalize search paths
Attempt to canonicalize search paths, then fall back to using absolute
path. This preserves the fix for FACT-1510, but also ensures we
deduplicate most paths.

This primarily fixes that in Puppet we load `lib/facter` from both
Puppet's search paths and Facter's internal handling of Ruby's LOAD_PATH.
Previously canonicalization would ensure that the resulting file paths for
custom facts were identical, so when we check whether we've previously
loaded a file we would identify duplicates and avoid loading them again.
The fix for FACT-1510 broke that. This is the least-impact approach to
fixing it.

Also use forward-slash explicitly when working with paths that come from
Ruby, to avoid inconsistent representations when we can't determine the
canonical path.