(FACT-1810) Fix open timeout for fetching ec2 metadata
The open_timeout wasn't specified, which means that it spent 60s
for each call to open (9 times in total). With this extra argument,
the run duration for facter goes down from over 9 minutes to
several seconds.

A regular 'curl 169.254.169.254' also hangs for a long time on
those machines, unless the --connect-timeout switch is specified.