(FACT-1597) Improve error handling when querying AIX disk devices
Previously, Facter would attempt to stumble forward with disk
information on AIX, even when it was impossible to query. We now
treat errors from the OS as an indication that a device probably
is not actually a disk, and report a debug message instead of
trying to populate the fact anyway.

Hopefully this will resolve all the issues of "unkown disk type"
being thrown with nonsense values for the device type field.