(FACT-1475) Clear out bogus IPV4 data for IPV6 query
On some platforms, in particular CentOS 5 and 6, facter erroneously
outputs ipv4 values for ipv6 facts when ipv6 is disabled. Specifically,
it outputs an ipv4 address for the ip6 fact. The reason it does this is
because the command `ip -6 route show` returns ipv4 data. Facter uses
this command to figure out the primary network interface and to also
provide at least the network interface's ip address when getifaddr does
not return any bindings for it. Facter fills in the ip6 fact when ipv6
is disabled for exactly this reason -- getifaddr does not return any
ipv6 bindings, but `ip -6 route show` outputs a bogus ipv4 address that
gets filled in instead.

This commit detects when the ip command outputs ipv4 information for an
ipv6 query, correctly erasing all currently collected ipv6 route data when
this occurs.