(FACT-1497) Try calculating the uptime from the utmpx file first
Previously, we would parse the output of the `uptime` command to
get the system uptime in our POSIX resolver. Unfortunately, there
is a bug in the `uptime` command that causes the seconds entry to
update itself once the minutes entry is updated. Thus, invoking the
`uptime` command a few seconds after the previous invocation would
still show the same uptime value, which is unexpected behavior.

Our AIX and Solaris platforms rely on the POSIX resolver to determine
their uptime information. Thus, the bug in the `uptime` command meant
that the `system_uptime` fact on these platforms was not being updated
correctly.

This commit tries to calculate the uptime programmatically before
calling the `uptime` command. It does this by attempting to get the system
boot time from the utmpx file, then subtracting that value from the current
system time. If it fails to read the boot time from the utmpx file, then
it delegates to getting the uptime information from the `uptime` command
itself.

The new method's been verified to work on our AIX and Solaris platforms.