(FACT-1550) Get os.architecture from processors.models[0] on AIX
On AIX, we can dynamically add and remove resources from the LPAR.
This includes removing proc0, which is what we originally checked
to obtain the OS architecture. In this case, we will report an
empty value for the OS architecture, which is incorrect.

This commit uses the processors.model[0] fact to obtain the
architecture instead, where the processors fact directly queries the
ODM for all of the available processors on the system and thus is not
affected by any LPAR configuration.

We also clean up the code here a bit by using our internal ODM APIs
instead of shelling out to lsattr and parsing its output to get the
hardware info. It is better to use our internal ODM APIs instead to
keep things consistent with our other AIX fact resolvers and to
avoid shelling out when we don't need to. The resulting code's
also a lot cleaner.