(FACT-1525) Add pages_to_bytes helper in AIX memory resolver
Previously, we would compute the number of bytes from a given number
of pages as `<num_pages> * PAGE_SIZE`. The problem is that if `<num_pages>`
is not a 64 bit numeric type, then the expression's value would be evaluated
as a uint so that large byte values would overflow. This would cause
Facter to report overflowed values for the swapsize and swapsize_mb facts
since the `size` and `free` fields of the pginfo struct are uints.

We could fix this by just casting these uints to uint64_t in the
multiplication. It's cleaner, however, to just add a helper that
converts pages to bytes to encapsulate the casting and the conversion.