(FACT-1873) Use builtin fact if custom facts don't resolve or have 0 wt.
Previously, if any builtin facts have conflicting custom facts that all
fail to resolve, and at least one of the custom facts has a weight > 0,
Facter will not report a value for that fact. This is undesirable
behavior because we always want to fall back to the builtin value if
all custom facts fail to resolve. Note that the "weight > 0" part is
important because Facter _will_ use the builtin value if all custom
facts have 0 weight.

Furthermore, the builtin fact should have precedence over 0 weight
custom facts. The existing code enforces this condition when _all_
custom facts have 0 weight, but it does _not_ enforce the condition
when only _some_ custom facts have 0 weight. In the latter case, the 0
weight custom facts will take precedence over the builtin fact value.

This commit fixes the fact resolution logic so that it falls back to the
built-in fact if all custom facts fail to resolve, while also ensuring
that it has precedence over 0 weight custom facts.