(FACT-1900) Only check for `sysnative` if we are running under Wow64
Previously, facter would check for the existance of
`%sysroot%\sysnative` and assume it was the correct system32 directory
if it exists. This is incorrect if that directory has somehow been
created as a normal directory in the normal filesystem.

Instead, we should just check for whether we are running under WOW64,
and spit out the correct directory based on that query. We don't use
IsWow64Process2, since we need to support versions of Windows which do
not have it available.