(FACT-1910) Remove check that mount point is physical
This removes the check that the returned mount points are physical
devices or temporary filesystems, and instead will return mount points
for all mounts on the system. Removing this check was the original
suggestion for FACT-1519 to allow tmpfs mount points, but instead was
replaced with a flag to allow only tmpfs filesystems. Instead of adding
additional flags for different filesystems to allow, I propose removing
this check entirely. This will expands the data collected by facter and
allow more fine grained system configuration.

Skip over any mount for /proc or /sys

The puppetlabs facter team said that facter should not return the cgroup
and kernel-internal filesystems. This adds a check to skip the mount
point if it is mounted under the proc or sys directories. It keeps the
tmpfs check to prevent breaking any code that looks for the
currently-returned tmpfs mount information under these directories.

Test the filesystem resolver itself

This adds checks for the filesystem resolver to add filesystem,
mountpoint, and partition facts. Additionally it checks the logic of the
collect_mountpoint_data function to ensure that the expected mount types
will be populated by facter.

Remove check for tmpfs mount under /proc

This check was added to ensure that existing data returned by facter
about mounts under /proc and /sys remained unchanged. However these
mounts vary between different machines and kernel revision. At least for
now I think it is best to remove this check instead of trying to pin a
specific mount under /proc or /sys.