(FACT-1914) Expose noop Facter.reset method in JRuby
The PAL layer calls `Facter.reset`, so this method is needed for
compatibility with PAL. We leave it unimplemented, since facts are
treated as immutable in JRuby and custom facts are ignored.

If we ever do need to properly reset, we will need to figure out how
to properly implement thread safety and environment isolation for the
Facter java bindings.