(FACT-1916): fix route parsing on Linux
This PR removes the assumption that a every all values in the out put of
ip route show includes key value pairs.  !1283 first introduced this
assumption fixing.  This PR also removes the patch for FACT-1405 introduced
in !1372 as it is no longer needed.  one could further remove the code
from !1464 to address FACT-1394 as that seems to be cause by the same
issue, however i have left that in as one dosen't need to parse linkdown
routes.

Currently the none cases of values which are not key value pairs are
 * flags: linkdown, pervasive, onlink
 * params with optional values:  mtu [lock] value

 We could add something like the following to the loop at
https://github.com/puppetlabs/facter/blob/master/lib/src/facts/linux/networking_resolver.cc#L198
however this may be overkill as the flags and mtuy alwasy come after the
dev and src paramaters.

```c++
unordered_set<string> route_flags {
  "linkdown",
  "pervasive",
  "onlink"
}
size_t step = 2;
for (size_t i = dst_idx+1; i < parts.size(); i += step) {
    std::string key(parts[i].begin(), parts[i].end());
    std::string value(parts[i+1].begin(), parts[i+1].end());
    step = 2;
    if route_flags.find(key) {
      step = 1;
      continue;
    } else if (key == "mtu" and value == "lock"){
      step = 3;
      continue;
    }
    if (key == "dev") {
        r.interface.assign(value);
    }
    if (key == "src") {
        r.source.assign(value);
    }
}
```

I also wonder if its worth always passing `-d` to `ip route show` to
ensure the route_type is always included but i'm unsure if that flag is
available on all linux