(FACT-1929) Add serial_number to the disks fact
Make it possible include a serial_number as part of a disk fact.

On Linux systems, the serial number is not exposed in /sys/devices, and
is not accessible through blkid.  lsblk(1) rely on udev to get the
information so do likely in facter: add an optional dependency on
libudev and use it when available.

On FreeBSD, the information is directly available from GEOM.